% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{get_samples}
\alias{get_samples}
\title{get_samples}
\usage{
get_samples(df, num_samples)
}
\arguments{
\item{df}{A data frame of webppl "histogram" output}

\item{num_samples}{A number of samples to reconstruct}
}
\description{
Turn webppl "histogram" output into samples.
}
\examples{
get_samples(df, 1000)
}

