% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{install_webppl_package}
\alias{install_webppl_package}
\title{Install webppl package}
\usage{
install_webppl_package(package_name, path = global_pkg_path())
}
\arguments{
\item{package_name}{Name of package to be installed}

\item{path}{Path to package install location (defaults to webppl's global
package directory)}
}
\description{
Install an npm package to webppl's global installation.
}
\examples{
\dontrun{install_webppl_package("babyparse")}
}

