% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{kill_webppl}
\alias{kill_webppl}
\title{Kill rwebppl processes}
\usage{
kill_webppl(pids = NULL)
}
\arguments{
\item{pids}{(optional) Vector of process IDs to kill (defaults to killing all
rwebppl processes)}
}
\description{
Kill rwebppl processes
}
\examples{
\dontrun{kill_webppl()}
\dontrun{kill_webppl(6939)}
}
