% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{link_webppl}
\alias{link_webppl}
\title{Symlink global webppl install to rwebppl directory}
\usage{
link_webppl(existingLoc = NULL)
}
\arguments{
\item{path}{Path to global webppl installation (defaults to npm root)}
}
\description{
If you installed webppl with rwebppl and later decided to install it globally, 
it's useful to replace the rwebppl install with a symlink to the global install
}
\examples{
\dontrun{link_webppl()}
}

