% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{uninstall_webppl_package}
\alias{uninstall_webppl_package}
\title{Uninstall webppl package}
\usage{
uninstall_webppl_package(package_name, path = global_pkg_path())
}
\arguments{
\item{package_name}{Name of package to be installed}

\item{path}{Path to package install location (defaults to webppl's global
package directory)}
}
\description{
Uninstall an npm package from webppl's global installation.
}
\examples{
\dontrun{uninstall_webppl_package("babyparse")}
}
