% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{upgrade_webppl}
\alias{upgrade_webppl}
\title{Upgrade webppl installation}
\usage{
upgrade_webppl()
}
\description{
Upgrades local (or symlinked) webppl installation to newest version allowed
by rwebppl (currently 0.8.1)
}
\examples{
\dontrun{upgrade_webppl()}
}

