% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{get_samples}
\alias{get_samples}
\title{Get samples}
\usage{
get_samples(df, num_samples)
}
\arguments{
\item{df}{A data frame of webppl "histogram" output (has a column called
`prob` with probabilities, remaining columns are parameter values).}

\item{num_samples}{A number of samples to reconstruct.}
}
\value{
Data frame of parameter values with number of rows equal to
  `num_samples`.
}
\description{
Turn webppl "histogram" output into samples.
}
\examples{
program <- "
  var model = function() {
    var theta = uniform(0, 1)
    var x = flip(theta)
    return x
  }
"
num_samples <- 100
df <- webppl(program_code = program, model_var = "model",
             inference_opts = list(method = "MCMC", samples = num_samples))
get_samples(df, num_samples)
}

