% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{run_webppl}
\alias{run_webppl}
\title{webppl}
\usage{
run_webppl(program_code = NULL, program_file = NULL, data = NULL,
  data_var = "data", packages = NULL, model_var = "model",
  inference_opts = NULL, chains = NULL, random_seed = NULL,
  sort_by = "prob", chain = 1)
}
\arguments{
\item{program_code}{A string of a webppl program.}

\item{program_file}{A file containing a webppl program.}

\item{data}{A data frame (or other serializable object) that can be
referenced in the program.}

\item{data_var}{A name by which data can be referenced in the program.}

\item{packages}{A character vector of external package names to use.}

\item{model_var}{The name by which the model be referenced in the program.}

\item{inference_opts}{Options for inference
(see http://webppl.readthedocs.io/en/master/inference.html)}

\item{chains}{Number of chains (this run is one chain).}

\item{random_seed}{Seed for random number generator}

\item{sort_by}{Sort probability table by probability or support (enumeration only)}

\item{chain}{Chain number of this run.}
}
\description{
Runs a webppl program.
}
