% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{webppl}
\alias{webppl}
\title{webppl}
\usage{
webppl(program_code = NULL, program_file = NULL, data = NULL,
  data_var = NULL, packages = NULL, model_var = NULL,
  inference_opts = NULL, chains = 1, cores = 1,
  output_format = "webppl")
}
\arguments{
\item{program_code}{A string of a webppl program.}

\item{program_file}{A file containing a webppl program.}

\item{data}{A data frame (or other serializable object) that can be
referenced in the program.}

\item{data_var}{A name by which data can be referenced in the program.}

\item{packages}{A character vector of external package names to use.}

\item{model_var}{The name by which the model be referenced in the program.}

\item{inference_opts}{Options for inference
(see http://webppl.readthedocs.io/en/master/inference.html)}

\item{chains}{Number of times to run the program (defaults to 1).}

\item{cores}{Number of cores to use when running multiple chains (defaults to
1).}

\item{output_format}{An optional string indicating posterior output format:
"webppl" probability table (default), "samples" for just the samples, 
"ggmcmc" for use with ggmcmc package.}
}
\value{
The program's return value(s).
}
\description{
Runs a webppl program.
}
\examples{
program_code <- "flip(0.5)"
webppl(program_code)
}

