% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwebppl.R
\name{webppl}
\alias{webppl}
\title{webppl}
\usage{
webppl(model_code = NULL, model_file = NULL, model_packages = NULL)
}
\arguments{
\item{model_code}{A string of a webppl program.}

\item{model_file}{A file containing a webppl program.}

\item{model_packages}{A character vector of external package names to use.}
}
\value{
The model's return value(s).
}
\description{
Runs webppl model.
}
\examples{
model_code <- "flip(0.5)"
webppl(model_code)
}

